refactored cli based on iac changes
