+++ unilake-cli/src/Unilake.Cli/Args/UpOptions.cs
@@ -1,4 +1,5 @@
 using CommandLine;
+using Unilake.Cli.Stacks;
 using Parser = Unilake.Cli.Config.Parser;
 
 namespace Unilake.Cli.Args;
