+++ unilake-cli/src/Unilake.Cli/Config/IConfigNode.cs
@@ -3,7 +3,6 @@
 public interface IConfigNode
 {
     string Section { get; }
-    // TODO: 1 add params so you can specify which items are required
     IEnumerable<ValidateResult> Validate(EnvironmentConfig config, IConfigNode? parentNode, params string[] checkProps);
     public static bool CheckProp(string propName, string[] checkProps) =>
         checkProps.Length == 0 || (!checkProps.Contains(string.Empty) && checkProps.Contains(propName));
