+++ unilake-iac/src/Unilake.Iac/Unilake.Iac.csproj
@@ -7,8 +7,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Pulumi" Version="3.55.1" />
-      <PackageReference Include="Pulumi.Kubernetes" Version="3.30.2" />
+      <PackageReference Include="Pulumi" Version="3.58.0" />
+      <PackageReference Include="Pulumi.Kubernetes" Version="4.5.4" />
     </ItemGroup>
 
     <ItemGroup>
